refactor: move auth context/hooks from _lib to _entities
- Move auth-context.tsx, use-auth.ts, use-logout.ts, user-storage.ts
  from _lib/auth/ to _entities/auth/
- Remove circular dependency: lib no longer imports from entities
- Update ESLint boundaries config to enforce unidirectional imports
- Delete _lib/query-keys.ts (merged into auth-context.tsx)
- Update all imports to use new paths
- Update CLAUDE.md and README.md documentation
