+++ app/_entities/auth/index.ts
@@ -5,3 +5,9 @@ export type {
   ProfileType,
 } from './model';
 export { login, logout } from './mutations';
+export { AuthProvider, AuthContext } from './auth-context';
+export type { AuthContextValue, AuthUser } from './auth-context';
+export { useAuth } from './use-auth';
+export { useLogout } from './use-logout';
+export type { StoredUser, GetStoredUserResult } from './user-storage';
+export { UserStorageError } from './user-storage';
