--- app/_lib/query-keys.ts
@@ -1,3 +1,0 @@
-import { PET_PUBLIC_QUERY_KEYS } from '@/app/_entities/pet/query-keys';
-
-export const PUBLIC_QUERY_KEYS = [...PET_PUBLIC_QUERY_KEYS] as const;
