+++ app/auth/login/_components/login-form.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { AuthProvider } from '@/app/_lib/auth/auth-context';
+import { AuthProvider } from '@/app/_entities/auth';
 import LoginForm from './login-form';
 
 const mockPush = jest.fn();
